requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
yfinance>=0.2.28
python-dotenv>=1.0.0
pandas>=2.0.0
lxml>=4.9.0
python-dateutil>=2.8.2
flask>=2.3.0

